- Add CI pipeline - Fix Tests Tests do not depend on configuration anymore (which they don't need yet) - Refactor DB.ts and split up in to multiple Repositories. - Changed config loading behaviour to prevent loading on file initilization.
