+++ src/WebServer.ts
@@ -1,6 +1,6 @@
 import express, { Express } from "express";
 import * as http from "http";
-import { configuration } from "./config/Config";
+import { getConfig } from "./config/Config";
 import { log } from "./Util";
 
 // Create a new express app instance
@@ -21,7 +21,7 @@ export class WebServer {
     }
 
     public start() {
-        let httpConfig = configuration.get().http;
+        let httpConfig = getConfig().get().http;
         this.server = this.app.listen(parseInt(httpConfig.port), httpConfig.host, function () {
             log("info",`Web Server is listening on ${httpConfig.host}:${httpConfig.port} .`);
         });
