+++ src/commands/DiscordLog.ts
@@ -1,6 +1,6 @@
 import * as discord from "discord.js";
-import { log } from "../Util";
 import { BotgartCommand } from "../BotgartCommand";
+import { log } from "../Util";
 
 /**
 Testcases:
@@ -32,7 +32,7 @@ export class DiscordLog extends BotgartCommand {
 
     command(message, responsible, guild, args) {
         const cl = this.getBotgartClient();
-        cl.db.addLogChannel(guild, args.type, args.channel);
+        cl.logChannelRepository.addLogChannel(guild, args.type, args.channel);
         log("notice", "Set up log channel '{0}' for event type '{1}' in guild '{2}'.".formatUnicorn(args.channel.name, args.type, guild.name));
         (<discord.Message>message).react("✅"); // that's a white checkmark, even if not rendered properly...
     }
