+++ src/commands/FindDuplicates.ts
@@ -1,7 +1,7 @@
-import { log } from "../Util";
-import * as L from "../Locale";
 import * as discord from "discord.js";
 import { BotgartCommand } from "../BotgartCommand";
+import * as L from "../Locale";
+import { log } from "../Util";
 
 export class FindDuplicates extends BotgartCommand {
     constructor() {
@@ -14,7 +14,7 @@ export class FindDuplicates extends BotgartCommand {
 
     command(message: discord.Message, responsible: discord.User, guild: discord.Guild, args: any): void {
         const cl = this.getBotgartClient();
-        cl.db.findDuplicateRegistrations().forEach(d => {
+        cl.registrationRepository.findDuplicateRegistrations().forEach(d => {
             // unknown users are already filtered out. Maybe we want to change that and notify the caller
             const userNames: string[] = d.users.split(",");
             Promise.all(userNames.map(async u => await guild.members.fetch(u)).filter(u => u))
