+++ src/commands/faq/DeleteFAQ.ts
@@ -1,7 +1,6 @@
-import * as L from "../../Locale";
 import * as discord from "discord.js";
-import { BotgartClient } from "../../BotgartClient";
 import { BotgartCommand } from "../../BotgartCommand";
+import * as L from "../../Locale";
 
 /**
 Testcases:
@@ -30,7 +29,7 @@ export class DeleteFaq extends BotgartCommand {
     }
 
     command(message: discord.Message , responsible: discord.User, guild: discord.Guild, args: any): boolean {
-        let deleted: boolean = (<BotgartClient>this.client).db.deleteFAQ(args.key, guild.id);
+        let deleted: boolean = this.getBotgartClient().faqRepository.deleteFAQ(args.key, guild.id);
         let reply: string = deleted ? L.get("FAQ_DELETED").formatUnicorn(args.key) : L.get("FAQ_NOT_DELETED").formatUnicorn(args.key);
         this.reply(message, responsible, reply)
         return deleted;
