+++ src/commands/permanentrole/PermanentRoleAdd.ts
@@ -1,8 +1,8 @@
-import * as Util from "../../Util";
-import * as L from "../../Locale";
 import * as discord from "discord.js";
 import { BotgartClient } from "../../BotgartClient";
 import { BotgartCommand } from "../../BotgartCommand";
+import * as L from "../../Locale";
+import * as Util from "../../Util";
 
 /**
 Testcases:
@@ -36,7 +36,7 @@ export class AddPermanentRole extends BotgartCommand {
         }
 
         let cl = <BotgartClient>this.client;
-        let success = cl.db.storePermanentRole(args.member.user.id, message.guild.id, args.role.name);
+        let success = cl.permanentRoleRepository.storePermanentRole(args.member.user.id, message.guild.id, args.role.name);
 
         if(success) {
             Util.log("info", "Successfully added role {0} to user {0} in guild {0}.".formatUnicorn(args.role.name, args.member.user.username, message.guild.name));
