+++ src/commands/permanentrole/PermanentRoleRemove.ts
@@ -1,8 +1,8 @@
-import * as Util from "../../Util";
-import * as L from "../../Locale";
 import * as discord from "discord.js";
 import { BotgartClient } from "../../BotgartClient";
 import { BotgartCommand } from "../../BotgartCommand";
+import * as L from "../../Locale";
+import * as Util from "../../Util";
 
 /**
 Testcases:
@@ -36,7 +36,7 @@ export class RemovePermanentRole extends BotgartCommand {
         }
 
         let cl = <BotgartClient>this.client;
-        let success = cl.db.deletePermanentRole(args.member.user.id, message.guild.id, args.role.name);
+        let success = cl.permanentRoleRepository.deletePermanentRole(args.member.user.id, message.guild.id, args.role.name);
 
         if(success) {
             Util.log("info", "Successfully removed role {0} from user {0} in guild {0}.".formatUnicorn(args.role.name, args.member.user.username, message.guild.name));
